+++ mod/wikirate_csv_import/set/type/relationship_answer_import_file.rb
@@ -19,7 +19,7 @@ COLUMNS = {
   comment: "Comment"
 }
 
-def csv_row_class
+def import_item_class
   ImportItem::Structure::RelationshipAnswerCsv
 end
