+++ mod/wikirate_csv_import/spec/set/type/source_import_file_spec.rb
@@ -8,7 +8,7 @@ RSpec.describe Card::Set::Type::SourceImportFile do
   end
 
   include_context "csv import" do
-    let(:csv_row_class) { ImportItem::Structure::SourceCsv }
+    let(:import_item_class) { ImportItem::Structure::SourceCsv }
     let(:import_card) { Card["source import test"] }
     let(:data) do
       {
